requirements.txt

streamlit
opencv-python
numpy
psycopg2-binary
mlflow
